Lumping the one test formerly in APIUtilTest (?) into the SanitizerTest
